// migrations/snapshots/a9819f75a74bcb335ad4cfb5a2d1e44f1b6a2f34842b7b611080b10237325db7/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Problem: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'];
      readonly difficulty: 'EASY' | 'MEDIUM' | 'HARD';
      readonly constraints: CodecTypes['pg/text@1']['output'];
      readonly examples: CodecTypes['pg/json@1']['output'];
      readonly supportedLanguages: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly TestCase: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly problemId: CodecTypes['pg/text@1']['output'];
      readonly input: CodecTypes['pg/text@1']['output'];
      readonly expectedOutput: CodecTypes['pg/text@1']['output'];
      readonly isHidden: CodecTypes['pg/bool@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Problem: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'];
      readonly difficulty: 'EASY' | 'MEDIUM' | 'HARD';
      readonly constraints: CodecTypes['pg/text@1']['input'];
      readonly examples: CodecTypes['pg/json@1']['input'];
      readonly supportedLanguages: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly TestCase: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly problemId: CodecTypes['pg/text@1']['input'];
      readonly input: CodecTypes['pg/text@1']['input'];
      readonly expectedOutput: CodecTypes['pg/text@1']['input'];
      readonly isHidden: CodecTypes['pg/bool@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly problem: {
      readonly constraints: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'];
      readonly difficulty: 'EASY' | 'MEDIUM' | 'HARD';
      readonly examples: CodecTypes['pg/json@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly supportedLanguages: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly testCase: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly expectedOutput: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly input: CodecTypes['pg/text@1']['output'];
      readonly isHidden: CodecTypes['pg/bool@1']['output'];
      readonly problemId: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly problem: {
      readonly constraints: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'];
      readonly difficulty: 'EASY' | 'MEDIUM' | 'HARD';
      readonly examples: CodecTypes['pg/json@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly supportedLanguages: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly testCase: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly expectedOutput: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly input: CodecTypes['pg/text@1']['input'];
      readonly isHidden: CodecTypes['pg/bool@1']['input'];
      readonly problemId: CodecTypes['pg/text@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;